LayoutTests/inspector/codemirror/prettyprinting-javascript-expected.txt

Test JavaScript pretty printing with CodeMirror.


== Running test suite: CodeMirror.PrettyPrinting.JavaScript
-- Running test case: CodeMirror.PrettyPrinting.JavaScript.block-comment.js
PASS

-- Running test case: CodeMirror.PrettyPrinting.JavaScript.single-statement-blocks.js
PASS

-- Running test case: CodeMirror.PrettyPrinting.JavaScript.switch-case-default.js
PASS

-- Running test case: CodeMirror.PrettyPrinting.JavaScript.unary-binary-operators.js
PASS

-- Running test case: CodeMirror.PrettyPrinting.JavaScript.do-while-loop.js
PASS

-- Running test case: CodeMirror.PrettyPrinting.JavaScript.if-followed-by-while.js
PASS

-- Running test case: CodeMirror.PrettyPrinting.JavaScript.while-within-do-while.js
PASS

-- Running test case: CodeMirror.PrettyPrinting.JavaScript.if-while-within-do-while.js
PASS

-- Running test case: CodeMirror.PrettyPrinting.JavaScript.do-while-within-if.js
PASS
